seed data provinsi

+++ migrations/20220329063517-create-provinsi.js
@@ -12,10 +12,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       latitude: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT(11,10)
       },
       longitude: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT(11,10)
       },
       createdAt: {
         allowNull: false,
